refactor(fe): organize theme preset list
