+++ frontend/core/unit/DashboardThread/Appearance/Theme/PresetList/index.tsx
@@ -1,86 +1,15 @@
-import { AnimatePresence, domAnimation, LazyMotion, m, useAnimationControls } from 'motion/react'
-import type { Transition } from 'motion/react'
-import { useEffect, useState } from 'react'
+import { AnimatePresence, domAnimation, LazyMotion, m } from 'motion/react'
 
 import { THEME_PRESET, THEME_PRESET_OPTIONS } from '~/const/theme_preset'
 import useTrans from '~/hooks/useTrans'
 import ArrowsSplitSVG from '~/icons/ArrowsSplit'
 
+import useSalon from '../salon/preset_list'
+import type { TThemePresetCardMode, TThemePresetOption } from '../spec'
+import { CARD_LAYOUT_TRANSITION, cardVariants, listVariants, ROTATE_ANGLES } from './constant'
 import PresetCard from './PresetCard'
-import useSalon, { ROTATE_ANGLES } from './salon/preset_list'
-import type { TThemePresetCardMode, TThemePresetOption, TThemePresetOverwrite } from './spec'
-
-const CARD_LAYOUT_TRANSITION = {
-  duration: 0.8,
-  ease: [0.22, 1, 0.36, 1],
-} as const
-
-const CARD_STAGGER = 0.1
-
-const INCOMING_DOT_TRANSITION: Transition = {
-  duration: 0.72,
-  ease: 'linear',
-  times: [0, 0.73, 1],
-}
-
-const OUTGOING_TOP_DOT_TRANSITION: Transition = {
-  duration: 2.3,
-  ease: [0.22, 1, 0.36, 1],
-  times: [0, 0.84, 1],
-}
-
-const OUTGOING_BOTTOM_DOT_TRANSITION: Transition = {
-  duration: 2.3,
-  ease: [0.22, 1, 0.36, 1],
-  delay: 0.56,
-  times: [0, 0.84, 1],
-}
-
-const listVariants = {
-  visible: (direction: number) => ({
-    transition: {
-      delayChildren: direction > 0 ? 0.02 : 0,
-      staggerChildren: CARD_STAGGER,
-      staggerDirection: direction,
-    },
-  }),
-}
-
-const cardVariants = {
-  hidden: (direction: number) => ({
-    opacity: 0,
-    x: 28 * direction,
-    scale: 0.92,
-  }),
-  visible: {
-    opacity: 1,
-    x: 0,
-    scale: 1,
-  },
-  exit: (direction: number) => ({
-    opacity: 0,
-    x: 36 * direction,
-    scale: 0.9,
-  }),
-}
-
-type TPresetListItem =
-  | {
-      type: 'preset'
-      preset: TThemePresetOption
-    }
-  | {
-      type: 'forkedFrom'
-    }
-
-type TProps = {
-  activePreset: string
-  activePresetBase: string
-  hasCustomPreset: boolean
-  customOverwrite: TThemePresetOverwrite
-  showForkRelation: boolean
-  onSelect: (preset: TThemePresetOption) => void
-}
+import type { TPresetListItem, TProps } from './spec'
+import useForkDotsAnimation from './useForkDotsAnimation'
 
 export default function PresetList({
   activePreset,
@@ -92,10 +21,8 @@ export default function PresetList({
 }: TProps) {
   const s = useSalon({ showForkRelation })
   const { t } = useTrans()
-  const [hoveredPreset, setHoveredPreset] = useState<string | null>(null)
-  const incomingDotControls = useAnimationControls()
-  const outgoingTopDotControls = useAnimationControls()
-  const outgoingBottomDotControls = useAnimationControls()
+  const { incomingDotControls, outgoingTopDotControls, outgoingBottomDotControls } =
+    useForkDotsAnimation(showForkRelation)
   const basePreset =
     THEME_PRESET_OPTIONS.find((preset) => preset.value === activePresetBase) ??
     THEME_PRESET_OPTIONS[0]
@@ -118,56 +45,6 @@ export default function PresetList({
         { type: 'preset', preset: basePreset },
       ]
     : presetOptions.map((preset) => ({ type: 'preset' as const, preset }))
-
-  useEffect(() => {
-    if (!showForkRelation) return
-
-    let live = true
-    let animationTimer: ReturnType<typeof setTimeout> | null = null
-
-    const playForkDots = async () => {
-      incomingDotControls.set({ x: 64, y: 0, opacity: 0, scale: 1 })
-      outgoingTopDotControls.set({ x: -2, y: -4, opacity: 0, scale: 0 })
-      outgoingBottomDotControls.set({ x: -2, y: 4, opacity: 0, scale: 0 })
-
-      await incomingDotControls.start({
-        x: [64, 2, 2],
-        y: [0, 0, 0],
-        opacity: [0.9, 0.85, 0],
-        scale: [1, 0.3, 0],
-        transition: INCOMING_DOT_TRANSITION,
-      })
-
-      if (!live) return
-
-      void outgoingTopDotControls.start({
-        x: [-2, -46, -46],
-        y: [-4, -4, -4],
-        opacity: [0.85, 0, 0],
-        scale: [0.35, 1, 1],
-        transition: OUTGOING_TOP_DOT_TRANSITION,
-      })
-      void outgoingBottomDotControls.start({
-        x: [-2, -48, -48],
-        y: [4, 4, 4],
-        opacity: [0.85, 0, 0],
-        scale: [0.35, 1, 1],
-        transition: OUTGOING_BOTTOM_DOT_TRANSITION,
-      })
-    }
-
-    animationTimer = setTimeout(() => {
-      void playForkDots()
-    }, CARD_LAYOUT_TRANSITION.duration * 1000)
-
-    return () => {
-      live = false
-      if (animationTimer) clearTimeout(animationTimer)
-      incomingDotControls.stop()
-      outgoingTopDotControls.stop()
-      outgoingBottomDotControls.stop()
-    }
-  }, [incomingDotControls, outgoingBottomDotControls, outgoingTopDotControls, showForkRelation])
 
   return (
     <LazyMotion features={domAnimation}>
@@ -241,12 +118,8 @@ export default function PresetList({
                 <PresetCard
                   preset={preset}
                   active={active}
-                  activeSuppressed={
-                    active && hoveredPreset !== null && hoveredPreset !== preset.value
-                  }
                   mode={cardMode}
                   rotateAngle={showForkRelation ? 0 : (ROTATE_ANGLES[index] ?? 0)}
-                  onHover={setHoveredPreset}
                   onSelect={onSelect}
                 />
               </m.div>
