+++ frontend/core/unit/DashboardThread/Appearance/Theme/salon/preset_list/rotate.ts
